requirements.txt
# Runtime dependencies for 0DTE bot and backtesting
numpy
pandas
scipy
yfinance
alpaca-py
pydantic
tenacity
python-dotenv
schedule
watchdog

